Create README.md
BookDazzle is a platform that offers summaries of the top 15 books across various genres. Users can explore book summaries and have the option to purchase the full versions directly from the platform. It features secure user authentication, ensuring that user IDs and passwords are safely stored in the database. The platform is optimized for easy navigation, delivering a smooth and user-friendly experience for readers. Overall, it combines convenience, security, and simplicity to enhance book discovery and purchasing.